Conflicts:
	README.md
	asm/macros/event.inc
	include/config/general.h
	include/field_player_avatar.h
	src/cable_car.c
	src/data.c
	src/data/item_icon_table.h
	src/event_object_movement.c
	src/field_player_avatar.c
	src/naming_screen.c
	src/overworld.c
	src/scrcmd.c

+++ src/cable_car.c
@@ -795,7 +795,7 @@ static void CreateCableCarSprites(void)
     //! Always use normal state
     u16 playerGraphicsId = GetPlayerAvatarGraphicsIdByStateIdAndGender(PLAYER_AVATAR_STATE_NORMAL, gSaveBlock2Ptr->playerGender);
     u16 rval = Random();
-    u8 hikerGraphicsIds[4] = {
+    u16 hikerGraphicsIds[4] = {
         OBJ_EVENT_GFX_HIKER,
         OBJ_EVENT_GFX_CAMPER,
         OBJ_EVENT_GFX_PICNICKER,
@@ -1058,4 +1058,3 @@ static void InitGroundTilemapData(bool8 goingDown)
 
     sCableCar->groundTimer = 0;
 }
-
